pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scientific-ml-autoresearch"
version = "0.1.0"
description = "A minimal autonomous research workflow for scientific machine learning."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Yongsheng Chen" }
]
license = { text = "MIT" }
dependencies = [
  "typer>=0.12",
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "jinja2>=3.1",
  "rich>=13.7"
]

[project.scripts]
autoresearch = "autoresearch.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
